requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.28.35
botocore==1.31.35
jmespath==1.0.1
s3transfer==0.6.2
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.19
django-allauth==0.54.0
django-appconf==1.0.5
django-storages==1.13.2
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
Pillow==10.0.0
psycopg2-binary==2.9.6
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.16
